Compilation warning..... -_-

+++ Classes/GON_NSString+Utils.m
@@ -26,8 +26,8 @@
 {
     NSMutableString *output = [NSMutableString string];
     const unsigned char* source = (const unsigned char *)[self UTF8String];
-    int sourceLen = (int)strlen((const char *)source);
-    for (int i = 0; i < sourceLen; ++i) {
+    NSUInteger sourceLen = strlen((const char *)source);
+    for (NSUInteger i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
             [output appendString:@"+"];
